move notifier API to V1
