+++ jormungandr/src/rest/v1/handlers.rs
@@ -1,4 +1,4 @@
-use crate::rest::{v1::logic, ContextLock};
+use crate::rest::{context::Error, v1::logic, ContextLock};
 use warp::{reject::Reject, Rejection, Reply};
 
 impl Reject for logic::Error {}
@@ -37,4 +37,19 @@ pub async fn get_fragment_logs(context: ContextLock) -> Result<impl Reply, Rejec
         .await
         .map_err(warp::reject::custom)
         .map(|r| warp::reply::json(&r))
+}
+
+pub async fn handle_subscription(
+    ws: warp::ws::Ws,
+    context: ContextLock,
+) -> Result<impl Reply, Rejection> {
+    let context = context.read().await;
+    logic::handle_subscription(ws, &context)
+        .await
+        .map_err(warp::reject::custom)
+    // let full_context = context.try_full().map_err(warp::reject::custom)?;
+
+    // let notifier: crate::notifier::Notifier = full_context.notifier.clone();
+
+    // Ok(ws.on_upgrade(move |socket| add_connection(notifier, socket)))
 }
