+++ jormungandr/src/rest/v1/logic.rs
@@ -115,3 +115,17 @@ pub async fn get_fragment_logs(context: &Context) -> Result<Vec<FragmentLog>, Er
     .instrument(span)
     .await
 }
+
+pub async fn handle_subscription(
+    ws: warp::ws::Ws,
+    context: &Context,
+) -> Result<impl warp::Reply, Error> {
+    let full_context = context.try_full()?;
+    let notifier: crate::notifier::Notifier = full_context.notifier.clone();
+
+    Ok(ws.on_upgrade(move |socket| add_connection(notifier, socket)))
+}
+
+async fn add_connection(notifier: crate::notifier::Notifier, socket: warp::ws::WebSocket) {
+    notifier.new_connection(socket).await;
+}
